kojis controller fixed
