+++ app/assets/stylesheets/components/_index_toggler.scss
@@ -1,0 +1,10 @@
+.toggle-button {
+  width: 225px;
+  height: 50px;
+  margin-left:auto;
+  margin-right:auto;
+  margin-top: 20px;
+  margin-bottom: 20px;
+  background-color: $beige;
+  border-radius: 30px;
+}
