+++ app/views/foods/index.html.erb
@@ -1,4 +1,4 @@
-<div class="index-container">
+<div class="index-container" data-controller="kojis-toggler">
   <div class="bar-container">
     <div class="bar-item active" id="favorites-tab">Favorites</div>
     <div class="bar-item" id="history-tab">History</div>
