+++ db/schema.rb
@@ -86,9 +86,9 @@ ActiveRecord::Schema[7.1].define(version: 2024_03_05_025701) do
     t.datetime "remember_created_at"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.boolean "admin", default: false, null: false
     t.string "username"
     t.string "preferences"
-    t.boolean "admin", default: false, null: false
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
